+++ ChangeLog.md
@@ -5,9 +5,8 @@ All major/minor changes between releases will be documented here.
 ## Roadmap: 1.2.64 - 1.2.69
 
 - Create an instance-specific `/about/{instance} page that will render just the site info that is typically in the sidebar
-- Revamp `/profile` section:
-    - Move `/saved` route from root to `/profile/saved`
-    - Create layout to Keep user info sidebar on all `/profile` pages.
+   
+    
 
 ## 1.2.65
 ### Bugfixes
@@ -15,6 +14,9 @@ All major/minor changes between releases will be documented here.
 
 ### Enhancements
 - Vimeo embeds can now autoplay in posts if you have that option enabled (default is off)
+- "Saved" posts are now under your profile section
+- "Inbox" is now under the profile section.
+- All sub-pages in the profile section now have a common layout with your user info in the sidebar